Extract DeploymentOrder class
